+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Patch] Font-weight 400 for tiny button
 - [Patch] Added auto-focus to the Textarea
 - [Patch] Added delay and hideDelay prop for the poptip