+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 - [Patch] Added auto-focus to the Textarea
 - [Patch] Added delay and hideDelay prop for the poptip
 - [Feature] Support Buttons that look like Links and Links that look like Buttons
+- [Feature] Create Sidebar component
 
 ## 40.4.1 - 2018-10-01
 - [Patch] Refine Switch disabled styles